demoLangGraph/requirements.txt
langgraph
langchain-core
langchain-google-genai
google-generativeai
python-dotenv
langgraph-cli[inmem]
redis
booktype
psutil